# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiFieldError": {
        "additionalProperties": false,
        "description": "One problem found in a benchmark file, as the API reports it.\n\nDeliberately NOT :class:`~llm_eval_lab.models.FieldError`, and deliberately\ntwo fields shorter than it.\n\n``file`` is dropped because it is the absolute path of the file on the\nserver, and an error response is not the place to describe the server's file\nsystem to a client.\n\n``input_repr`` is dropped because it is a repr of whatever the loader was\nlooking at when validation failed, and that is benchmark-file content. When\na required field is missing, pydantic reports the failure at the location of\nthe FIELD and echoes the whole PARENT object as the input - so a case that\nomits ``id`` and carries a credential in its ``metadata`` echoed the\ncredential back, in a body the API returned with a 200. Redacting by\ninspecting the location name could not see that, because the location named\nthe missing field and not the secret beside it. Removing the member removes\nthe class of defect, and matches what\n:func:`request_validation_errors` already does for request bodies. The\nlocation, the case id and the message say what is wrong; the file on the\noperator's disk says what it is wrong in.",
        "properties": {
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "location",
          "case_id",
          "message"
        ],
        "title": "ApiFieldError",
        "type": "object"
      },
...
      "CaseDelta": {
        "additionalProperties": false,
        "description": "How one case moved between the baseline run and the candidate run.",
        "properties": {
          "baseline_passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Passed"
          },
          "baseline_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Score"
          },
          "candidate_passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Passed"
          },
          "candidate_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Score"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          }
        },
        "required": [
          "case_id",
          "baseline_score",
          "candidate_score",
          "delta",
          "baseline_passed",
          "candidate_passed"
        ],
        "title": "CaseDelta",
        "type": "object"
      },
...
      "CheckOutcome": {
        "description": "The result of one threshold check, with its advisory statistics.\n\n``p_value`` is the McNemar exact test and is populated only for paired\npass-rate comparisons. ``confidence_interval`` is the Newcombe Method-10\ninterval on the paired pass-rate delta, and is ``None`` for every other\nmetric and in unpaired mode, where the two Wilson intervals in\n``baseline_ci`` and ``candidate_ci`` take its place instead.",
        "properties": {
          "baseline": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline"
          },
          "baseline_ci": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Ci"
          },
          "candidate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate"
          },
          "candidate_ci": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Ci"
          },
          "confidence_interval": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Interval"
          },
          "delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          },
          "direction": {
            "$ref": "#/components/schemas/ThresholdDirection"
          },
          "intervals_overlap": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intervals Overlap"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "n_baseline": {
            "title": "N Baseline",
            "type": "integer"
          },
          "n_candidate": {
            "title": "N Candidate",
            "type": "integer"
          },
          "n_paired": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Paired"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "p_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P Value"
          },
          "relative_delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative Delta"
          },
          "significant": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Significant"
          },
          "status": {
            "$ref": "#/components/schemas/CheckStatus"
          },
          "threshold_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold Value"
          },
          "violated_bound": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Violated Bound"
          }
        },
        "required": [
          "metric",
          "label",
          "direction",
          "status",
          "baseline",
          "candidate",
          "delta",
          "relative_delta",
          "violated_bound",
          "threshold_value",
          "n_baseline",
          "n_candidate",
          "n_paired",
          "confidence_interval",
          "p_value",
          "significant",
          "note"
        ],
        "title": "CheckOutcome",
        "type": "object"
      },
      "CheckStatus": {
        "description": "Outcome of a single regression threshold check.",
        "enum": [
          "passed",
          "failed",
          "warning",
          "insufficient_data",
          "missing_metric"
        ],
        "title": "CheckStatus",
        "type": "string"
      },
      "CompareRequest": {
        "additionalProperties": false,
        "description": "Body of ``POST /api/compare``.\n\n``thresholds`` is an inline policy, not a path: the API accepts no\nserver-side file paths. Omitting it uses the policy shipped with the package,\nwhich is the same default ``llm-eval compare`` uses.",
        "properties": {
          "baseline_run_id": {
            "$ref": "#/components/schemas/ShortText"
          },
          "candidate_run_id": {
            "$ref": "#/components/schemas/ShortText"
          },
          "thresholds": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RegressionThresholds"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "baseline_run_id",
          "candidate_run_id"
        ],
        "title": "CompareRequest",
        "type": "object"
      },
...
      "MetricCheck": {
        "additionalProperties": false,
        "description": "One threshold applied to one metric. Pure data; it evaluates nothing.\n\nBound semantics, which the phase 4 engine in ``reporting/regression.py``\nimplements and which are carried in the field descriptions so there is one\nwritten source: ``min_value`` is an INCLUSIVE floor, violated when the\ncandidate is strictly below it. ``max_value`` is an EXCLUSIVE ceiling,\nviolated when the candidate is greater than or equal to it. That asymmetry\nis what makes \"error rate must remain below 1%\" expressible as\n``max_value: 0.01``, where a candidate of exactly ``0.01`` fails.\n\nEvaluation deliberately does not live here. Gate semantics are engine\nbehaviour, not contract shape, and a frozen model carrying them would force\nevery later correction through the frozen-file protocol.",
        "properties": {
          "direction": {
            "$ref": "#/components/schemas/ThresholdDirection",
            "description": "Which way this metric has to move to be an improvement. Descriptive: it labels the metric for rendering and arrow direction. The bounds below are already directional, so an engine must not derive a second directional rule from this field."
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "max_absolute_decrease": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Absolute Decrease"
          },
          "max_absolute_increase": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Absolute Increase"
          },
          "max_relative_decrease": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Relative Decrease"
          },
          "max_relative_increase": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Relative Increase"
          },
          "max_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "EXCLUSIVE ceiling. Violated when the candidate is greater than or equal to it, so a candidate exactly equal to max_value FAILS. That asymmetry is what makes 'error rate must remain below 1%' expressible as max_value: 0.01.",
            "title": "Max Value"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "min_samples": {
            "default": 1,
            "title": "Min Samples",
            "type": "integer"
          },
          "min_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "INCLUSIVE floor. Violated when the candidate is strictly below it, so a candidate exactly equal to min_value passes.",
            "title": "Min Value"
          },
          "require_significant": {
            "default": false,
            "title": "Require Significant",
            "type": "boolean"
          },
          "severity": {
            "default": "error",
            "enum": [
              "error",
              "warning"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "metric",
          "direction"
        ],
        "title": "MetricCheck",
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "An RFC 9457 problem detail, with this API's extension members.\n\nEvery extension is optional and omitted when absent, so a client can read\n``status`` and ``title`` from any error and reach for the rest only when it\nknows the ``type``.",
        "properties": {
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "env_var": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Env Var"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ApiFieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "extra": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "RegressionReport": {
        "description": "A complete baseline-versus-candidate comparison.\n\n``mode`` is ``paired`` when the two runs share case ids and ``unpaired``\notherwise. Thresholds evaluate point estimates in both modes, so the gate\nverdict never depends on the mode.",
        "properties": {
          "baseline_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Label"
          },
          "baseline_run_id": {
            "title": "Baseline Run Id",
            "type": "string"
          },
          "baseline_suite_hash": {
            "title": "Baseline Suite Hash",
            "type": "string"
          },
          "candidate_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Label"
          },
          "candidate_run_id": {
            "title": "Candidate Run Id",
            "type": "string"
          },
          "candidate_suite_hash": {
            "title": "Candidate Suite Hash",
            "type": "string"
          },
          "changed_cases": {
            "items": {
              "type": "string"
            },
            "title": "Changed Cases",
            "type": "array"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/CheckOutcome"
            },
            "title": "Checks",
            "type": "array"
          },
          "comparable": {
            "title": "Comparable",
            "type": "boolean"
          },
          "generated_at": {
            "$ref": "#/components/schemas/UtcDatetime"
          },
          "incomparable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incomparable Reason"
          },
          "intervals_overlap": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intervals Overlap"
          },
          "mode": {
            "enum": [
              "paired",
              "unpaired"
            ],
            "title": "Mode",
            "type": "string"
          },
          "only_in_baseline": {
            "items": {
              "type": "string"
            },
            "title": "Only In Baseline",
            "type": "array"
          },
          "only_in_candidate": {
            "items": {
              "type": "string"
            },
            "title": "Only In Candidate",
            "type": "array"
          },
          "paired_case_count": {
            "title": "Paired Case Count",
            "type": "integer"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "suite_hash_match": {
            "title": "Suite Hash Match",
            "type": "boolean"
          },
          "summary": {
            "$ref": "#/components/schemas/RegressionSummary"
          },
          "thresholds_id": {
            "title": "Thresholds Id",
            "type": "string"
          },
          "thresholds_version": {
            "title": "Thresholds Version",
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          }
        },
        "required": [
          "generated_at",
          "baseline_run_id",
          "candidate_run_id",
          "baseline_label",
          "candidate_label",
          "verdict",
          "mode",
          "comparable",
          "incomparable_reason",
          "suite_hash_match",
          "baseline_suite_hash",
          "candidate_suite_hash",
          "paired_case_count",
          "only_in_baseline",
          "only_in_candidate",
          "changed_cases",
          "thresholds_id",
          "thresholds_version",
          "checks",
          "summary"
        ],
        "title": "RegressionReport",
        "type": "object"
      },
      "RegressionSummary": {
        "additionalProperties": false,
        "description": "Which cases flipped, and the largest score drops behind the headline.",
        "properties": {
          "largest_score_drops": {
            "items": {
              "$ref": "#/components/schemas/CaseDelta"
            },
            "title": "Largest Score Drops",
            "type": "array"
          },
          "n_flipped": {
            "title": "N Flipped",
            "type": "integer"
          },
          "newly_failing": {
            "items": {
              "type": "string"
            },
            "title": "Newly Failing",
            "type": "array"
          },
          "newly_passing": {
            "items": {
              "type": "string"
            },
            "title": "Newly Passing",
            "type": "array"
          },
          "still_failing": {
            "items": {
              "type": "string"
            },
            "title": "Still Failing",
            "type": "array"
          }
        },
        "required": [
          "newly_failing",
          "newly_passing",
          "still_failing",
          "n_flipped",
          "largest_score_drops"
        ],
        "title": "RegressionSummary",
        "type": "object"
      },
      "RegressionThresholds": {
        "additionalProperties": false,
        "description": "A named, versioned threshold policy.",
        "properties": {
          "checks": {
            "items": {
              "$ref": "#/components/schemas/MetricCheck"
            },
            "minItems": 1,
            "title": "Checks",
            "type": "array"
          },
          "confidence_level": {
            "default": 0.95,
            "title": "Confidence Level",
            "type": "number"
          },
          "fail_on_missing_metric": {
            "default": true,
            "title": "Fail On Missing Metric",
            "type": "boolean"
          },
          "id": {
            "default": "default",
            "title": "Id",
            "type": "string"
          },
          "min_paired_cases": {
            "default": 1,
            "title": "Min Paired Cases",
            "type": "integer"
          },
          "paired": {
            "default": true,
            "title": "Paired",
            "type": "boolean"
          },
          "require_same_suite": {
            "default": true,
            "title": "Require Same Suite",
            "type": "boolean"
          },
          "version": {
            "default": "1",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "checks"
        ],
        "title": "RegressionThresholds",
        "type": "object"
      },
...
      "ShortText": {
        "maxLength": 200,
        "minLength": 1,
        "type": "string"
      },
...
      "ThresholdDirection": {
        "description": "Which way a metric has to move for the change to be an improvement.",
        "enum": [
          "higher_is_better",
          "lower_is_better"
        ],
        "title": "ThresholdDirection",
        "type": "string"
      },
...
      "UtcDatetime": {
        "format": "date-time",
        "type": "string"
      },
...
      "Verdict": {
        "description": "Overall outcome of a regression comparison, and the CI exit-code source.",
        "enum": [
          "pass",
          "warn",
          "fail",
          "incomparable"
        ],
        "title": "Verdict",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Local-first LLM evaluation: benchmarks, runs, statistics and regression gating. Errors are RFC 9457 problem details with the media type application/problem+json.",
    "title": "llm-eval-lab",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/compare": {
      "get": {
        "description": "Compare two runs under the shipped default policy.\n\nThe linkable form, for a dashboard that wants a comparison in a URL. A custom\npolicy needs a body, so it needs the ``POST``.",
        "operationId": "compare_by_query_api_compare_get",
        "parameters": [
          {
            "in": "query",
            "name": "baseline_run_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Baseline Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "candidate_run_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Candidate Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegressionReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "A bearer token is required and was absent or wrong."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No run, case or benchmark suite has that identifier."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The request conflicts with the current state of the run."
          },
          "411": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "A request body was sent without declaring its length."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The request body exceeds the configured cap."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The request or the benchmark suite it names did not validate."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "An unmapped failure, reported with a correlation id and nothing else."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The database could not be reached."
          }
        },
        "summary": "Compare By Query",
        "tags": [
          "compare"
        ]
      },
      "post": {
        "description": "Compare a candidate run against a baseline under an inline or default policy.",
        "operationId": "compare_api_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegressionReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "A bearer token is required and was absent or wrong."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No run, case or benchmark suite has that identifier."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The request conflicts with the current state of the run."
          },
          "411": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "A request body was sent without declaring its length."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The request body exceeds the configured cap."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The request or the benchmark suite it names did not validate."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "An unmapped failure, reported with a correlation id and nothing else."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The database could not be reached."
          }
        },
        "summary": "Compare",
        "tags": [
          "compare"
        ]
      }
    },
...
  }
}